+++ Cargo.toml
@@ -17,6 +17,7 @@ fast_image_resize = "0.7.0"
 
 serde_derive = "1.0.133"
 serde = "1.0.133"
+num_enum = "0.5.6"
 
 [profile.release]
 debug = true
